[FIX]: Seperate scene file structure
