+++ src/phaser/scenes/multiplay/Preloader.js
@@ -1,0 +1,7 @@
+import SinglePreloader from "../singleplay/Preloader";
+
+export default class MultiPreloader extends SinglePreloader {
+  completeLoading() {
+    this.scene.start("multi");
+  }
+}
