+++ src/phaser/scenes/multiplay/index.js
@@ -1,0 +1,24 @@
+import Phaser from "phaser";
+
+import Preloader from "./Preloader";
+import Stage from "./Stage";
+
+const config = {
+  type: Phaser.WEBGL,
+  scale: {
+    mode: Phaser.Scale.FIT,
+    parent: "game-container",
+    autoCenter: Phaser.Scale.CENTER_BOTH,
+  },
+  backgroundColor: "#60E7F7",
+  physics: {
+    default: "arcade",
+    arcade: {
+      gravity: { y: 0 },
+      debug: true,
+    },
+  },
+  scene: [Preloader, Stage],
+};
+
+export default config;
