+++ src/phaser/scenes/singleplay/Preloader.js
@@ -1,7 +1,7 @@
 import Phaser from "phaser";
 
-import createHeroAnimations from "../animations/Hero";
-import createEnemyAnimations from "../animations/Enemy";
+import createHeroAnimations from "../../animations/Hero";
+import createEnemyAnimations from "../../animations/Enemy";
 
 export default class Preloader extends Phaser.Scene {
   preload() {
@@ -89,7 +89,7 @@ export default class Preloader extends Phaser.Scene {
     this.load.image("tiles", "iso-12-tileset.png");
     this.load.image("coin", "coin.png");
     this.load.image("add-coin", "add-coin.png");
-    
+
     this.load.image("start", "start.png");
     this.load.image("win", "win.png");
     this.load.image("lose", "lose.png");
