+++ src/phaser/scenes/singleplay/Stage1.js
@@ -1,34 +1,39 @@
-import Hero from "../gameObjects/Hero";
-import Enemy from "../gameObjects/Enemy";
-import ChaseHeroAI from "../ai/ChaseHeroAI";
+import Hero from "../../gameObjects/Hero";
+import Enemy from "../../gameObjects/Enemy";
+import ChaseHeroAI from "../../ai/ChaseHeroAI";
 
-import store from "../../store";
-import { updateGameProgress } from "../../redux/slices/singlePlaySlice";
-import { gameProgress } from "../../constants/gameState";
-import { Level3 } from "../../constants/enemyList";
+import store from "../../../store";
+import { updateGameProgress } from "../../../redux/slices/singlePlaySlice";
+import { gameProgress } from "../../../constants/gameState";
+import { Level1 } from "../../../constants/enemyList";
+import { ThemeProvider } from "styled-components";
 
-export default class Stage3 extends Phaser.Scene {
+export default class Stage1 extends Phaser.Scene {
   constructor() {
-    super("stage3");
+    super("stage1");
   }
 
   init() {
     this.cameras.main.fadeIn(500, 0, 0, 0);
+
+    this.registry.values.score = 0; // 변수
+    this.registry.values.time = 0; // 변수
   }
 
   create() {
-    this.setBackground(3);
+    this.game.events.emit("gameStart"); // 변수
+
+    this.setBackground(1);
 
     this.setStatusBar();
 
-    this.setCharacters(Level3);
+    this.setCharacters(Level1);
 
     this.setCoinToMap();
     
     this.cursors = this.input.keyboard.createCursorKeys();
 
-    this.cameras.main.startFollow(this.hero, true);
-    this.cameras.main.setZoom(1);
+    this.setCamera();
   }
 
   update(time, delta) {
@@ -107,51 +112,19 @@ export default class Stage3 extends Phaser.Scene {
   }
 
   setCharacters(enemyList) {
-    this.hero = new Hero(this, 100, 550, "hero");
-
-    this.add.existing(this.hero).setDepth(5);
-
-    this.physics.world.enable(
-      this.hero,
-      Phaser.Physics.Arcade.DYNAMIC_BODY
-    );
-
-    this.hero.body.setSize(40, 110, true);
+    this.createHero();
     
-    this.enemies = enemyList.map((enemy) => {
-      const newEnemy = new Enemy(this, enemy.x, enemy.y, "enemy");
-
-      newEnemy.setTargetIndicatorColor(enemy.indicatorColor);
-
-      this.add.existing(newEnemy).setDepth(5);
-
-      this.physics.world.enable(
-        newEnemy,
-        Phaser.Physics.Arcade.DYNAMIC_BODY
-      );
-      
-      newEnemy.body.setSize(40, 110, true);
-
-      const ai = enemy.ai;
-      
-      switch (ai) {
-        case "chase": {
-          newEnemy.setAI(new ChaseHeroAI(this.hero, newEnemy, this.boardLayer));
-          break;
-        }
-        default:
-          break;
-      }
-
-      return newEnemy;
-    });
-
-    this.physics.add.collider(this.enemies);
+    this.createEnemies(enemyList);
     
     this.physics.add.collider(this.hero, this.enemies, () => {
       this.moveNextStage();
       // this.stopStage();
     });
+  }
+
+  setCamera() {
+    this.cameras.main.startFollow(this.hero, true);
+    this.cameras.main.setZoom(1);
   }
 
   stopStage() {
@@ -183,7 +156,7 @@ export default class Stage3 extends Phaser.Scene {
         this.cameras.main.once(
           Phaser.Cameras.Scene2D.Events.FADE_OUT_COMPLETE,
           () => {
-            store.dispatch(updateGameProgress(gameProgress.GAME_CLEAR));
+            this.scene.start("stage2");
           }
         );
       },
@@ -212,5 +185,50 @@ export default class Stage3 extends Phaser.Scene {
     this.enemies.forEach((enemy) => {
       enemy.unSubscribeAI();
     });
+  }
+
+  createHero() {
+    this.hero = new Hero(this, 100, 550, "hero");
+
+    this.add.existing(this.hero).setDepth(5);
+
+    this.physics.world.enable(
+      this.hero,
+      Phaser.Physics.Arcade.DYNAMIC_BODY
+    );
+
+    this.hero.body.setSize(40, 110, true);
+  }
+
+  createEnemies(enemyList) {
+    this.enemies = enemyList.map((enemy) => {
+      const newEnemy = new Enemy(this, enemy.x, enemy.y, "enemy");
+
+      newEnemy.setTargetIndicatorColor(enemy.indicatorColor);
+
+      this.add.existing(newEnemy).setDepth(5);
+
+      this.physics.world.enable(
+        newEnemy,
+        Phaser.Physics.Arcade.DYNAMIC_BODY
+      );
+      
+      newEnemy.body.setSize(40, 110, true);
+
+      const ai = enemy.ai;
+      
+      switch (ai) {
+        case "chase": {
+          newEnemy.setAI(new ChaseHeroAI(this.hero, newEnemy, this.boardLayer));
+          break;
+        }
+        default:
+          break;
+      }
+
+      return newEnemy;
+    });
+
+    this.physics.add.collider(this.enemies);
   }
 }
