+++ src/phaser/scenes/singleplay/index.js
@@ -1,0 +1,26 @@
+import Phaser from "phaser";
+
+import Preloader from "./Preloader";
+import Stage1 from "./Stage1";
+import Stage2 from "./Stage2";
+import Stage3 from "./Stage3";
+
+const config = {
+  type: Phaser.WEBGL,
+  scale: {
+    mode: Phaser.Scale.FIT,
+    parent: "game-container",
+    autoCenter: Phaser.Scale.CENTER_BOTH,
+  },
+  backgroundColor: "#60E7F7",
+  physics: {
+    default: "arcade",
+    arcade: {
+      gravity: { y: 0 },
+      debug: true,
+    },
+  },
+  scene: [Preloader, Stage1, Stage2, Stage3],
+};
+
+export default config;
